Add more operation analysis, including deletes and inserts variables. And Using the preprocessor to decide the analysis being set or not.
The following are the new varialbes to record the opreations:

rtry  is the number of retries in the __list_find function.
cons  is the number of wait-free contains in the __list_find function that curr pointer pointed.
trav  is the number of list element traversal in the __list_find function.
fail  is the number of CAS() failures.
del   is the number of list_delete operation failed and restart again.
ins   is the number of list_insert operation failed and restart again.
load  is the number of atomic_load operation in list_delete, list_insert and __list_find.
store is the number of atomic_store operation in list_delete, list_insert and __list_find.